separate pairing from item
