+++ pairing.php
@@ -1,4 +1,5 @@
 <?php
+require "item.php";
 
 //TODO: Implement Glicko-2 algorithm and html generation functions
 class pairing
@@ -6,7 +7,7 @@ class pairing
     public string $id; //tournament id
     public item $p1; //item 1
     public item $p2; //item 2
-    public string $player;//the player name
+    public string $player; //the player name
     public int $winner; //I think ID would be best here, but 1/2 could work as well. bool in that case
 
     //TODO: get a pairing from 1 of 3 options: random, rating based, and reliability.
@@ -25,25 +26,4 @@ class pairing
     {
         throw new Error("sql not implemented");
     }
-}
-
-
-class item
-{
-    public int $id;
-    public string $name;
-    public int $rating;
-    public int $confidence;
-
-    //TODO: create constructor from PDO output
-    function __construct($in)
-    {
-        throw new Error("item constructor not implemented");
-    }
-
-    //TODO: generate sql to update the item in the database
-    function update(): string
-{
-        throw new Error("update not implemented");
-}
 }
